Moved _SOLVERS_INTERFACES to config as solvers
