+++ tests/test_cobrapy_compatibility.py
@@ -3,7 +3,6 @@ import nose
 from cobra.test import create_test_model
 from cobra.test.unit_tests import CobraTestCase, TestReactions
 from cobra.test.flux_analysis import TestCobraFluxAnalysis
-from cameo import load_model
 from cameo.solver_based_model import to_solver_based_model, SolverBasedModel
